src/Omnisearch_gpt/requirements.txt
Pillow==10.4.0
requests==2.32.3
google-search-results==2.4.2
serpapi==0.1.5


# File transfer and loading utilities
hf_transfer
dotenv

# Core ML frameworks
# vllm==0.7.3 # Do not uncomment this, update the vllm version in the Dockerfile
torch>=2.6,<2.7  # vllm==0.7.3 does not work with torch>=2.7
torchvision>=0.21.0
torchaudio
transformers
accelerate>=0.26.0
sentence-transformers


# Data processing and modeling
pandas
pydantic
datasets
jinja2>=3.1.0
tokenizers

# API integrations
openai
cragmm-search-pipeline>=0.5.0
chromadb>=1.0.0

# Misc
loguru
rich
tqdm
